Split weight proof helpers into utils module.
This keeps the handler focused while preserving behavior.
